+++ packages/firestore/src/core/pipeline-util.ts
@@ -29,24 +29,13 @@ import {
   gt,
   lte,
   gte,
-  Expr,
   eq,
   Or
 } from '../lite-api/expressions';
 import { Pipeline } from '../lite-api/pipeline';
 import { doc } from '../lite-api/reference';
 import { isNanValue, isNullValue } from '../model/values';
-import {
-  ArrayValue as ProtoArrayValue,
-  Function as ProtoFunction,
-  LatLng as ProtoLatLng,
-  MapValue as ProtoMapValue,
-  Pipeline as ProtoPipeline,
-  Timestamp as ProtoTimestamp,
-  Value as ProtoValue
-} from '../protos/firestore_proto_api';
 import { fail } from '../util/assert';
-import { isPlainObject } from '../util/input_validation';
 
 import { Bound } from './bound';
 import {
@@ -66,131 +55,6 @@ import {
 } from './query';
 
 /* eslint @typescript-eslint/no-explicit-any: 0 */
-
-function isITimestamp(obj: any): obj is ProtoTimestamp {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-  if (
-    'seconds' in obj &&
-    (obj.seconds === null ||
-      typeof obj.seconds === 'number' ||
-      typeof obj.seconds === 'string') &&
-    'nanos' in obj &&
-    (obj.nanos === null || typeof obj.nanos === 'number')
-  ) {
-    return true;
-  }
-
-  return false;
-}
-function isILatLng(obj: any): obj is ProtoLatLng {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-  if (
-    'latitude' in obj &&
-    (obj.latitude === null || typeof obj.latitude === 'number') &&
-    'longitude' in obj &&
-    (obj.longitude === null || typeof obj.longitude === 'number')
-  ) {
-    return true;
-  }
-
-  return false;
-}
-function isIArrayValue(obj: any): obj is ProtoArrayValue {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-  if ('values' in obj && (obj.values === null || Array.isArray(obj.values))) {
-    return true;
-  }
-
-  return false;
-}
-function isIMapValue(obj: any): obj is ProtoMapValue {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-  if ('fields' in obj && (obj.fields === null || isPlainObject(obj.fields))) {
-    return true;
-  }
-
-  return false;
-}
-function isIFunction(obj: any): obj is ProtoFunction {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-  if (
-    'name' in obj &&
-    (obj.name === null || typeof obj.name === 'string') &&
-    'args' in obj &&
-    (obj.args === null || Array.isArray(obj.args))
-  ) {
-    return true;
-  }
-
-  return false;
-}
-
-function isIPipeline(obj: any): obj is ProtoPipeline {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-  if ('stages' in obj && (obj.stages === null || Array.isArray(obj.stages))) {
-    return true;
-  }
-
-  return false;
-}
-
-export function isFirestoreValue(obj: any): obj is ProtoValue {
-  if (typeof obj !== 'object' || obj === null) {
-    return false; // Must be a non-null object
-  }
-
-  // Check optional properties and their types
-  if (
-    ('nullValue' in obj &&
-      (obj.nullValue === null || obj.nullValue === 'NULL_VALUE')) ||
-    ('booleanValue' in obj &&
-      (obj.booleanValue === null || typeof obj.booleanValue === 'boolean')) ||
-    ('integerValue' in obj &&
-      (obj.integerValue === null ||
-        typeof obj.integerValue === 'number' ||
-        typeof obj.integerValue === 'string')) ||
-    ('doubleValue' in obj &&
-      (obj.doubleValue === null || typeof obj.doubleValue === 'number')) ||
-    ('timestampValue' in obj &&
-      (obj.timestampValue === null || isITimestamp(obj.timestampValue))) ||
-    ('stringValue' in obj &&
-      (obj.stringValue === null || typeof obj.stringValue === 'string')) ||
-    ('bytesValue' in obj &&
-      (obj.bytesValue === null || obj.bytesValue instanceof Uint8Array)) ||
-    ('referenceValue' in obj &&
-      (obj.referenceValue === null ||
-        typeof obj.referenceValue === 'string')) ||
-    ('geoPointValue' in obj &&
-      (obj.geoPointValue === null || isILatLng(obj.geoPointValue))) ||
-    ('arrayValue' in obj &&
-      (obj.arrayValue === null || isIArrayValue(obj.arrayValue))) ||
-    ('mapValue' in obj &&
-      (obj.mapValue === null || isIMapValue(obj.mapValue))) ||
-    ('fieldReferenceValue' in obj &&
-      (obj.fieldReferenceValue === null ||
-        typeof obj.fieldReferenceValue === 'string')) ||
-    ('functionValue' in obj &&
-      (obj.functionValue === null || isIFunction(obj.functionValue))) ||
-    ('pipelineValue' in obj &&
-      (obj.pipelineValue === null || isIPipeline(obj.pipelineValue)))
-  ) {
-    return true;
-  }
-
-  return false;
-}
 
 export function toPipelineFilterCondition(f: FilterInternal): FilterCondition {
   if (f instanceof FieldFilterInternal) {
@@ -355,7 +219,7 @@ function whereConditionsFromCursor(
   bound: Bound,
   orderings: Ordering[],
   position: 'before' | 'after'
-): Expr {
+): FilterCondition {
   const cursors = bound.position.map(value => Constant._fromProto(value));
   const filterFunc = position === 'before' ? lt : gt;
   const filterInclusiveFunc = position === 'before' ? lte : gte;
